TransactionListFragment - solved problem when delete expense

+++ app/src/main/java/com/example/fico/presentation/helper/SwipeToDeleteCallback.kt
@@ -5,13 +5,14 @@ import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
 import com.example.fico.api.FormatValuesToDatabase
 import com.example.fico.model.Expense
-import com.example.fico.presentation.adapters.ExpenseListAdapter
+import com.example.fico.presentation.adapters.TransactionListAdapter
 import com.example.fico.presentation.viewmodel.TransactionListViewModel
+import com.example.fico.utils.constants.StringConstants
 
 class SwipeToDeleteCallback(
     private val recyclerView: RecyclerView,
     private val viewModel: TransactionListViewModel,
-    private val adapter: ExpenseListAdapter,
+    private val adapter: TransactionListAdapter,
     private val lifecycleOwner: LifecycleOwner
 ) :
     ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT) {
@@ -33,23 +34,28 @@ class SwipeToDeleteCallback(
 
         //Check if expense in not installment
         if (deleteItem.id.length < 41) {
-            val formattedPaymentDate = FormatValuesToDatabase().expenseDate(deleteItem.paymentDate)
-            val formattedPurchaseDate =
-                FormatValuesToDatabase().expenseDate(deleteItem.purchaseDate)
-            val formattedInputDateTime =
-                FormatValuesToDatabase().expenseDate(deleteItem.inputDateTime)
-            val expencePrice = "-${deleteItem.price.replace("R$ ", "").replace(",", ".")}"
-            val deleteItemFormatted = Expense(
-                deleteItem.id,
-                expencePrice,
-                deleteItem.description,
-                deleteItem.category,
-                formattedPaymentDate,
-                formattedPurchaseDate,
-                formattedInputDateTime
-            )
-            //Delete Item and update expense list
-            viewModel.deleteExpense(deleteItemFormatted)
+            if(deleteItem.type == StringConstants.DATABASE.EXPENSE){
+                val formattedPaymentDate = FormatValuesToDatabase().expenseDate(deleteItem.paymentDate)
+                val formattedPurchaseDate =
+                    FormatValuesToDatabase().expenseDate(deleteItem.purchaseDate)
+                val formattedInputDateTime =
+                    FormatValuesToDatabase().expenseDate(deleteItem.inputDateTime)
+                val expencePrice = "-${deleteItem.price.replace("R$ ", "").replace(",", ".")}"
+                val deleteItemFormatted = Expense(
+                    deleteItem.id,
+                    expencePrice,
+                    deleteItem.description,
+                    deleteItem.category,
+                    formattedPaymentDate,
+                    formattedPurchaseDate,
+                    formattedInputDateTime
+                )
+                //Delete Item and update expense list
+                viewModel.deleteExpense(deleteItemFormatted)
+            }else if(deleteItem.type == StringConstants.DATABASE.EARNING){
+                //TODO delete earning
+            }
+
         }else{
             viewModel.onInstallmentExpenseSwiped()
         }
